Rename app to workspace  in folder
