+++ packages/types/src/api/web/workspace/component.ts
