+++ packages/types/src/api/web/workspace/navigation.ts
