+++ packages/types/src/api/web/workspace/resource.ts
