+++ packages/types/src/api/web/workspace/static.ts
