+++ packages/types/src/api/web/workspace/view.ts
