+++ packages/types/src/api/web/workspace/workspaceFavourites.ts
